refactor func names
