+++ src/index.ts
@@ -1,8 +1,18 @@
 export { useScrycard } from "./hooks/useScrycard";
 export { default as Scrycard } from "./components/Scrycard";
-export { default as Scrytext } from "./components/Scrytext";
+export { default as Scryhover } from "./components/Scryhover";
 export { default as ScryNameCard } from "./components/ScryNameCard";
+export { default as Scrytext } from "./components/Scrytext";
 export {
     ScrycardsContextProvider,
     useScrycardsContext,
 } from "./contexts/scrycards";
+
+export { generateCardBorder, generateCardGradient } from "./utils/cardBorder";
+export { isColor, colorsFromCost } from "./utils/cardColors";
+export { cardTypesFromTypeline } from "./utils/cardType";
+export { fetchCards } from "./utils/fetchCards";
+export { fetchSymbols } from "./utils/fetchSymbols";
+export { matchCards } from "./utils/matchCards";
+
+export type * from "./types/scrycard";
